deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
numpy
scikit-learn==1.2.1
plotly
imblearn
feature_engine
joblib
streamlit_option_menu
pillow